Add AARCH64 channels for NixOS 20.09
This adds a channel for NixOS 20.09 on AArcch64 based on the hydra
jobset that we already have.

The topic of an aarch64 specific channel has come up regular and as
these devices are gaining in popularity will only be discussed more
often.

This had been discussed on the nixpkgs issue tracker: https://github.com/NixOS/nixpkgs/issues/83049